Add isArchivable to DatabaseRecord trait

+++ src/database/delete_macros.rs
@@ -5,7 +5,7 @@ macro_rules! delete_resource_where_fields {
         use pluralizer::pluralize;
         use time::OffsetDateTime;
         use anyhow::anyhow;
-
+        use crate::database::traits::DatabaseResource;
         async {
             let archived_at = OffsetDateTime::now_utc();
             let archived_at_str = archived_at.to_string();
@@ -13,12 +13,18 @@ macro_rules! delete_resource_where_fields {
             let resource_name = pluralize(&stringify!($resource).to_lowercase(), 2, false);
             let pool = get_connection().await;
 
-            let mut params = $params.clone();
-            params.push(("archived_at", &archived_at_str));
+            let params = $params.clone();
 
             let fields = params.iter().map(|field| field.0.to_string()).collect::<Vec<String>>();
             let values = params.iter().map(|field| field.1.to_string()).collect::<Vec<String>>();
-            let mut query = format!("DELETE FROM {} WHERE ", resource_name);
+
+            let mut query: String;
+            if <$resource as DatabaseResource>::is_archivable() {
+                query = format!("UPDATE {} SET archived_at = ${} WHERE ", resource_name, fields.len() + 1);
+            } else {
+                query = format!("DELETE FROM {} WHERE ", resource_name);
+            }
+
             for (i, field) in fields.iter().enumerate() {
                 query.push_str(&format!("{} = ${}", field, i + 1));
                 if i < fields.len() - 1 {
@@ -29,6 +35,9 @@ macro_rules! delete_resource_where_fields {
             let mut query = sqlx::query(&query);
             for (_, value) in values.iter().enumerate() {
                 query = query.bind(value);
+            }
+            if <$resource as DatabaseResource>::is_archivable() {
+                query = query.bind(&archived_at_str);
             }
 
             match query.execute(&pool).await {
